Converting .md to .html to create slide toggle test
I added a div class "drop-down" which will wrap all the content we have
and make it hidden by default in css.  Under each of those divs I will
make another div called(whatever the header's id is)-content.
Hopefully this works.  I converted the .md file to .html by clicking the
view source page on the browser and highlighting everything and putting
it in a new .html file and replaced the .md page with that new .html
file.